P1/PR-5: extract platform-profiles and content-scoring
@launchstack/tools/platform-profiles — one registry of what each social
platform expects, consolidating four scattered copies: posting
guidelines (run.ts), structure templates + few-shot examples
(generator.ts), hashtag caps (generation prompt), hard char limits
(publish.ts truncation), and the judge's calibration references
(benchmark/references/*.md → references.ts). The evaluate route and the
LLM-gated benchmark test now import the references instead of reading
files off disk, retiring the process.cwd() path guessing and the
outputFileTracingIncludes entry. MarketingPlatform vocabulary moves to
the tool; marketing re-exports it.

@launchstack/tools/content-scoring — one versioned rubric, two
consumers: the offline judge (scorePost, rubric moved verbatim,
temperature 0 via the tool models table) and the live quality gate
(validatePostQuality, still gated off in the live path). Benchmark
judges/ become re-export shims; the two existing benchmark test suites
pass unchanged. Variant ranking stays a P2, benchmark-gated change.

Verified: tools typecheck + 36 vitest tests green; features and web
typecheck clean; full web Jest suite green (174 suites / 2107 tests).
