+++ packages/features/src/marketing-pipeline/generator.ts
@@ -1,5 +1,7 @@
 import { z } from "zod";
 import { HumanMessage, SystemMessage } from "@langchain/core/messages";
+import { getPlatformProfile, type PlatformMeta } from "@launchstack/tools/platform-profiles";
+import { QUALITY_THRESHOLD, validatePostQuality } from "@launchstack/tools/content-scoring";
 import { invokeMarketingStructured } from "./models";
 import type {
     MarketingPlatform,
@@ -57,60 +59,6 @@ When a Messaging Strategy is provided:
 - Do NOT reference external companies or products from trend references as if they are your own features or proof points.
 - Keep the post aligned with the positioning angle while staying platform-native.`;
 
-function platformTemplate(platform: MarketingPlatform): string {
-    switch (platform) {
-        case "x":
-            return [
-                "Platform: X (Twitter)",
-                "Structure:",
-                "- Hook line first — a bold, concise claim or sharp observation. Front-load the insight.",
-                "- 1–2 high-signal follow-up lines that deliver concrete value or a surprising detail.",
-                "- Optional: 1 clear CTA or provocative question to drive replies.",
-                "- 0–2 relevant hashtags max. Skip them if they feel forced.",
-                "Constraints:",
-                "- Aim for ~280 characters. Brevity is the craft here—every word must pull weight.",
-                "- No thread format. This is a single, self-contained post.",
-                "Tone: Punchy, confident, conversational. Think founder tweet, not press release.",
-            ].join("\n");
-        case "linkedin":
-            return [
-                "Platform: LinkedIn",
-                "Structure:",
-                "- Line 1 (the hook): Use contrast, a counterintuitive claim, or a specific result.",
-                '  Good hooks: "Most teams do X. The ones winning do Y." / "We stopped doing X. Here\'s what happened."',
-                '  Bad hooks: "Excited to announce…" / "Introducing our new…"',
-                "- Body: Pick the format that best serves the content:",
-                "  Option A (Narrative): 3–6 short paragraphs telling a mini-story or walking through a shift in thinking.",
-                "  Option B (Educational breakdown): Structured sections with clear labels and bullet points to explain a concept, compare approaches, or present a framework.",
-                "  Either way, each section must deliver one clear insight. Weave in business impact naturally (time saved, risk reduced, clarity gained).",
-                "- Closing: End with a specific question that invites perspectives, or a takeaway the reader can act on.",
-                '  Good CTAs: "Where is your team on this?" / "What\'s working in your stack?"',
-                '  Bad CTAs: "Let\'s connect!" / "DM me for more info"',
-                "Tone: Professional but human. Think operator sharing a playbook, not company posting an ad.",
-            ].join("\n");
-        case "reddit":
-            return [
-                "Platform: Reddit",
-                "Structure:",
-                "- Open with a relatable pain point, a story, or a specific problem you encountered.",
-                "- Share what you learned, tried, or built — give real value (steps, a framework, a checklist).",
-                "- Keep the company mention minimal and natural. Never lead with it.",
-                "- Close with an honest, open-ended question that invites the community to share their experience.",
-                "Tone: Speak like a real person in the community, not a brand account.",
-                "Never use marketing-speak, CTAs, or promotional language. Redditors will call it out instantly.",
-            ].join("\n");
-        default:
-            return [
-                "Platform: General social",
-                "Structure:",
-                "- Lead with an insight or observation, not a product announcement.",
-                "- Deliver value in the body — a takeaway, a lesson, a useful framing.",
-                "- Close with a question or reflection that invites engagement.",
-                "Tone: Clear, conversational, value-first. Write like a person, not a brand.",
-            ].join("\n");
-    }
-}
-
 // helper to convert MarketingResearchResult[] into a compact text block
 function formatTrendReferences(research: MarketingResearchResult[]): string {
     if (!research.length) return "None available.";
@@ -133,94 +81,6 @@ function formatStrategyBlock(strategy: MessagingStrategy): string {
         `Human hook: ${strategy.humanHook}`,
         `Avoid: ${strategy.avoidList.join("; ")}`,
     ].join("\n");
-}
-
-const PLATFORM_EXAMPLES: Record<MarketingPlatform, string> = {
-    linkedin: `Two examples of strong LinkedIn posts (for style reference only — do NOT copy content):
-
-Example A — Narrative style (GOOD):
-"""
-Most marketing teams are still building campaigns the same way they did in 2019. The ones pulling ahead aren't just adopting AI — they're rethinking the entire pipeline.
-
-We spent the last quarter rebuilding how we go from insight to published content. The biggest shift wasn't the tools. It was accepting that manual review cycles were the bottleneck, not creative quality.
-
-Automated research cut our trend analysis from days to hours. Predictive targeting replaced our "spray and hope" approach with data-backed audience selection. And templated personalization let us run 4x the campaigns without scaling the team.
-
-The result isn't just speed — it's focus. Our team now spends time on strategy instead of spreadsheets.
-
-What's the biggest bottleneck in your marketing workflow right now?
-"""
-
-Example B — Educational breakdown style (GOOD):
-"""
-Marketing automation and marketing strategy are not the same thing. But they're increasingly being treated like they are.
-
-Here's the difference that matters:
-
-Marketing automation
-• Executes repetitive tasks at scale
-• Follows predefined rules and workflows
-• Optimizes what already exists
-
-Marketing strategy
-• Decides what to build and why
-• Adapts to market shifts and customer signals
-• Requires human judgment and context
-
-Where it gets interesting: AI is starting to bridge the gap.
-
-Predictive analytics can surface which segments are underserved. Trend analysis can flag shifts before they're obvious. And automated pipelines can test messaging variations faster than any team could manually.
-
-But the risk is real — when automation outpaces strategy, you're scaling the wrong things faster.
-
-Where does your team draw the line between automating and strategizing?
-"""
-
-Example C — BAD post (DO NOT write like this):
-"""
-Many teams find themselves buried under a mountain of documents, but those who succeed transform this challenge into an advantage.
-
-Here's how ProductX empowers you:
-- Feature A: It analyzes your documents, turning them into actionable insights.
-- Feature B: Our open-source platform benefits from a vibrant community.
-- Feature C: While enhancing customizability remains a priority, our workflows already support quick decision-making.
-
-This isn't just about speed—it's about clarity and focus.
-
-How does your team handle document overload? Where do you see the biggest opportunity to streamline?
-
-#Marketing #AI #Efficiency #Automation #Tech #ProductX
-"""
-Why this is bad: Generic opener, feature bullet list reads like a product page, acknowledges a weakness ("remains a priority"), two questions at the end, 6 hashtags is excessive, hollow claims without specific proof.`,
-    x: `Example of a strong X post (for style reference only — do NOT copy content):
-"""
-Your marketing team's bottleneck isn't creative talent — it's manual processes eating 60% of their week.
-
-We automated research + targeting and freed our team to actually think strategically.
-
-What's the one workflow you'd automate first?
-"""`,
-    reddit: `Example of a strong Reddit post (for style reference only — do NOT copy content):
-"""
-We were spending more time on campaign logistics than actual strategy. Trend research, audience segmentation, copy variations — all manual, all slow.
-
-So we built an internal pipeline that automates the repetitive parts. Not the creative decisions, just the grunt work: pulling trend data, matching it to our ICP, generating first drafts we can edit.
-
-Early results: campaigns go live in ~2 days instead of ~2 weeks. Quality is about the same (sometimes better, because we actually have time to think now).
-
-Curious if anyone else has tried automating parts of their marketing workflow. What worked, what didn't?
-"""`,
-    bluesky: `Example of a strong Bluesky post (for style reference only — do NOT copy content):
-"""
-Hot take: the biggest AI opportunity in marketing isn't content generation — it's killing the busywork that keeps your team from doing actual strategy.
-
-We rebuilt our pipeline around that idea. Early results are promising.
-"""`,
-};
-
-interface PlatformMeta {
-    subreddit?: string;
-    hashtags?: string[];
 }
 
 function buildPrompt(args: {
@@ -263,11 +123,12 @@ function buildPrompt(args: {
         );
     }
 
+    const profile = getPlatformProfile(args.platform);
     parts.push(
         "",
-        platformTemplate(args.platform),
+        profile.structureTemplate,
         "",
-        PLATFORM_EXAMPLES[args.platform] ?? "",
+        profile.examples,
         "",
         "Task:",
         args.strategy
@@ -280,51 +141,13 @@ function buildPrompt(args: {
         "- NEVER acknowledge product weaknesses or areas under development.",
         "- Ground all product claims in the company context. Reframe anything unsupported as an industry observation.",
         "- End with exactly ONE question or CTA, not two.",
-        `- Use at most ${args.platform === "linkedin" ? 3 : args.platform === "x" ? 2 : 0} hashtags. Fewer is better.`,
+        `- Use at most ${profile.maxHashtags} hashtags. Fewer is better.`,
         args.platformMeta?.hashtags?.length
             ? "- Prefer the user's preferred hashtags over generic ones."
             : "",
         "- Return JSON matching the schema exactly."
     );
     return parts.join("\n");
-}
-
-/* ──────────────────────────────────────────────────────────────
- * Quality gate — optional post-generation validation
- * ────────────────────────────────────────────────────────────── */
-
-const QualityScoreSchema = z.object({
-    score: z.number().min(1).max(10),
-    issues: z.array(z.string()),
-    rewrite: z.string().nullable(),
-});
-
-const QUALITY_THRESHOLD = 6;
-
-async function validatePostQuality(
-    post: string,
-    platform: MarketingPlatform
-): Promise<{ score: number; issues: string[]; rewrite: string | null }> {
-    const response = await invokeMarketingStructured(
-        QualityScoreSchema,
-        [
-            new SystemMessage(
-                `You are a social media copy editor. Score this ${platform} post 1-10 on these criteria:
-1. Hook strength (does the first line stop the scroll?)
-2. Authenticity (does it sound like a person, not a brand?)
-3. Platform fit (does it match ${platform} conventions?)
-4. Specificity (are claims backed by concrete details, not vague hype?)
-5. Structure (is it narrative-driven rather than a feature list?)
-
-If score < ${QUALITY_THRESHOLD}, provide a "rewrite" field with an improved version that fixes the issues.
-Flag specific issues in "issues" array.`
-            ),
-            new HumanMessage(post),
-        ],
-        "quality_check"
-    );
-
-    return response;
 }
 
 /* ──────────────────────────────────────────────────────────────
