+++ packages/features/src/marketing-pipeline/run.ts
@@ -8,6 +8,7 @@ import { generateVariants } from "./generator";
 import { extractBrandVoice } from "@launchstack/tools/brand-voice";
 import { extractTargetPersona } from "@launchstack/tools/persona";
 import { checkClaimSources } from "@launchstack/tools/claim-evidence";
+import { getPlatformProfile } from "@launchstack/tools/platform-profiles";
 import { analyzeCompetitors } from "./competitor";
 import { buildMultiStrategy } from "./positioning";
 import {
@@ -207,8 +208,7 @@ export async function runMarketingPipeline(args: {
             const t3 = Date.now();
             emitStart("researching-trends", PG_GATHER);
             try {
-                const platformGuidelines = buildPlatformGuidelines(
-                    normalizedInput.platform,
+                const platformGuidelines = getPlatformProfile(normalizedInput.platform).guidelines(
                     normalizedInput.platformMeta
                 );
                 const basicContext = [
@@ -410,8 +410,7 @@ export async function runMarketingPipeline(args: {
     const t5 = Date.now();
     emitStart("generating-content");
 
-    const platformGuidelines = buildPlatformGuidelines(
-        normalizedInput.platform,
+    const platformGuidelines = getPlatformProfile(normalizedInput.platform).guidelines(
         normalizedInput.platformMeta
     );
     const companyContext = `${companyContextBase}\n\nPlatform best practices:\n${platformGuidelines}`;
@@ -538,47 +537,4 @@ export async function runMarketingPipeline(args: {
         },
         claimSources,
     };
-}
-
-function buildPlatformGuidelines(
-    platform: MarketingPipelineInput["platform"],
-    platformMeta?: MarketingPipelineInput["platformMeta"]
-): string {
-    switch (platform) {
-        case "reddit": {
-            const lines = [
-                "- Speak like a real community member, not a brand account.",
-                "- Lead with a specific pain point or story that matches the subreddit.",
-                "- Avoid pure self-promotion: focus on value, insight, or behind-the-scenes context.",
-                "- Use clear, descriptive titles; body can be longer and conversational.",
-                "- Invite discussion with an authentic question at the end.",
-            ];
-            if (platformMeta?.subreddit) {
-                lines.push(
-                    "",
-                    `Target subreddit: ${platformMeta.subreddit}`,
-                    "Tailor your tone, vocabulary, and content depth to match this subreddit's norms and audience expectations."
-                );
-            }
-            return lines.join("\n");
-        }
-        case "x":
-            return [
-                "- Keep posts tight and high-signal; front-load the hook in the first line.",
-                "- Use 1–2 sharp talking points instead of long paragraphs.",
-                "- Sprinkle in 1–2 relevant hashtags, but avoid hashtag spam.",
-                "- When appropriate, reference current trends or conversations in the space.",
-                "- Make the call-to-action explicit and easy to understand.",
-            ].join("\n");
-        case "linkedin":
-            return [
-                "- Use a strong first line that clearly states the outcome or insight.",
-                "- Write in short paragraphs or bullet points for easy scanning.",
-                "- Frame the message around business impact, transformation, or a concrete case.",
-                "- Keep the tone professional but human—less hype, more signal.",
-                "- Close with a takeaway or a soft call-to-action tailored to professionals.",
-            ].join("\n");
-        default:
-            return "- Write a clear, concise, value-focused message tailored to this platform.";
-    }
 }
