+++ packages/tools/src/platform-profiles/references.ts
@@ -1,0 +1,17 @@
+/**
+ * Curated reference posts per platform — the calibration examples the LLM
+ * judge scores against. Moved from benchmark/references/*.md (unification
+ * PR-5) so consumers import them instead of reading files off disk at
+ * runtime (which needed process.cwd() guessing and an outputFileTracingIncludes
+ * entry). Edit the strings here; bump JUDGE_RUBRIC_VERSION when calibration
+ * content changes meaningfully.
+ */
+
+export const REFERENCE_POSTS = {
+    x: "# X (Twitter) \u2014 Marketing Post Evaluation Reference\n\nFed **verbatim** to the LLM judge as the platform standard for X. The judge reads\nthe GOOD and BAD examples below \u2014 each with the company context it was written\nfrom \u2014 and uses them as the calibration bar when scoring a candidate post.\nThe judge SCORES ONLY; it never rewrites.\n\n> **Curation rules**: real, public\n> **marketing / advertisement** posts (a minor feature update promoted as an ad\n> counts) \u2014 **NOT** brand-new product-launch announcements. Target **10 good +\n> 3\u20135 bad** posts drawn from **>3 big companies** (AppLovin, Cursor, Perplexity,\n> General Translation, PostHog, \u2026). Keep excerpts short and attributed.\n\n## Company-context format (mirror of the generation pipeline)\n\nEach example carries the **same company-context window the generator sees**\n(`buildCompanyKnowledgeContext` / `buildMetadataContext` in\n[../../context.ts](../../context.ts)), so groundedness is judged against the same\nfacts the post was written from. Fill this block for every example:\n\n```text\nCompany Name:            # company table (companies)\nCompany Description:     # company table\nIndustry / Sector:       # company table\nEmployee Count Range:    # company table\nCompany Categories:      # category table (per company)\nKnowledge Base Signals:  # RAG over the company's document embeddings (top snippets)\n# Optional metadata block (present only when company_metadata JSONB exists):\n# === Services & Products ===\n# === Projects & Outcomes ===\n# === Markets ===\n```\n\n> **DB provenance** (matches what generation reads): identity fields come from the\n> `company` and `category` tables; _Knowledge Base Signals_ come from the RAG index\n> over the company's uploaded documents; the optional `=== \u2026 ===` sections come\n> from the `company_metadata` JSONB.\n\n## X platform norms\n\n<!-- fill: ~280-char focus, single-thought punch, thread norms, tone,\n     hashtag/emoji conventions, link placement -->\n\n---\n\n## GOOD examples\n\n### good-1\n\n**Company:** <!-- e.g. AppLovin -->\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's good:** <!-- one line -->\n\n### good-2\n\n**Company:**\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's good:**\n\n<!-- Repeat good-3 \u2026 good-10 (target 10, spread across >3 companies). -->\n\n---\n\n## BAD examples\n\n### bad-1\n\n**Company:**\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's bad:** <!-- one line: AI slop / feature-dump / off-platform tone / engagement-bait / ungrounded claim -->\n\n<!-- Repeat bad-2 \u2026 bad-5 (target 3\u20135). -->\n",
+    linkedin:
+        "# LinkedIn \u2014 Marketing Post Evaluation Reference\n\nFed **verbatim** to the LLM judge as the platform standard for LinkedIn. The judge\nreads the GOOD and BAD examples below \u2014 each with the company context it was\nwritten from \u2014 and uses them as the calibration bar when scoring a candidate post.\nThe judge SCORES ONLY; it never rewrites.\n\n> **Curation rules**: real, public\n> **marketing / advertisement** posts (a minor feature update promoted as an ad\n> counts) \u2014 **NOT** brand-new product-launch announcements. Target **10 good +\n> 3\u20135 bad** posts drawn from **>3 big companies** (AppLovin, Cursor, Perplexity,\n> General Translation, PostHog, \u2026). Keep excerpts short and attributed.\n\n## Company-context format (mirror of the generation pipeline)\n\nEach example carries the **same company-context window the generator sees**\n(`buildCompanyKnowledgeContext` / `buildMetadataContext` in\n[../../context.ts](../../context.ts)), so groundedness is judged against the same\nfacts the post was written from. Fill this block for every example:\n\n```text\nCompany Name:            # company table (companies)\nCompany Description:     # company table\nIndustry / Sector:       # company table\nEmployee Count Range:    # company table\nCompany Categories:      # category table (per company)\nKnowledge Base Signals:  # RAG over the company's document embeddings (top snippets)\n# Optional metadata block (present only when company_metadata JSONB exists):\n# === Services & Products ===\n# === Projects & Outcomes ===\n# === Markets ===\n```\n\n> **DB provenance** (matches what generation reads): identity fields come from the\n> `company` and `category` tables; _Knowledge Base Signals_ come from the RAG index\n> over the company's uploaded documents; the optional `=== \u2026 ===` sections come\n> from the `company_metadata` JSONB.\n\n## LinkedIn platform norms\n\n<!-- fill: longer-form professional tone, strong first line before the \"\u2026see more\"\n     fold, whitespace/line breaks for skimmability, 1\u20133 tasteful hashtags,\n     clear CTA, credibility over hype -->\n\n---\n\n## GOOD examples\n\n### good-1\n\n**Company:** <!-- e.g. PostHog -->\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's good:** <!-- one line -->\n\n### good-2\n\n**Company:**\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's good:**\n\n<!-- Repeat good-3 \u2026 good-10 (target 10, spread across >3 companies). -->\n\n---\n\n## BAD examples\n\n### bad-1\n\n**Company:**\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's bad:** <!-- one line: corporate clich\u00e9 / feature-dump / humble-brag with no substance / ungrounded claim -->\n\n<!-- Repeat bad-2 \u2026 bad-5 (target 3\u20135). -->\n",
+    reddit: "# Reddit \u2014 Marketing Post Evaluation Reference\n\nFed **verbatim** to the LLM judge as the platform standard for Reddit. The judge\nreads the GOOD and BAD examples below \u2014 each with the company context it was\nwritten from \u2014 and uses them as the calibration bar when scoring a candidate post.\nThe judge SCORES ONLY; it never rewrites.\n\n> **Curation rules**: real, public\n> **marketing / advertisement** posts (a minor feature update shared as an ad\n> counts) \u2014 **NOT** brand-new product-launch announcements. Target **10 good +\n> 3\u20135 bad** posts drawn from **>3 big companies** (AppLovin, Cursor, Perplexity,\n> General Translation, PostHog, \u2026). Keep excerpts short and attributed.\n> Reddit is community-first \u2014 the best \"marketing\" posts rarely read as ads.\n\n## Company-context format (mirror of the generation pipeline)\n\nEach example carries the **same company-context window the generator sees**\n(`buildCompanyKnowledgeContext` / `buildMetadataContext` in\n[../../context.ts](../../context.ts)), so groundedness is judged against the same\nfacts the post was written from. Fill this block for every example. For Reddit,\nalso note the **subreddit** (the generator takes `platformMeta.subreddit`):\n\n```text\nCompany Name:            # company table (companies)\nCompany Description:     # company table\nIndustry / Sector:       # company table\nEmployee Count Range:    # company table\nCompany Categories:      # category table (per company)\nKnowledge Base Signals:  # RAG over the company's document embeddings (top snippets)\nSubreddit:               # target community context (platformMeta.subreddit)\n# Optional metadata block (present only when company_metadata JSONB exists):\n# === Services & Products ===\n# === Projects & Outcomes ===\n# === Markets ===\n```\n\n> **DB provenance** (matches what generation reads): identity fields come from the\n> `company` and `category` tables; _Knowledge Base Signals_ come from the RAG index\n> over the company's uploaded documents; the optional `=== \u2026 ===` sections come\n> from the `company_metadata` JSONB.\n\n## Reddit platform norms\n\n<!-- fill: authentic community voice, no hashtags, no corporate-speak, value/story\n     first, discloses affiliation, fits the target subreddit's culture, invites\n     discussion rather than pitching -->\n\n---\n\n## GOOD examples\n\n### good-1\n\n**Company:** <!-- e.g. Cursor -->\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\nSubreddit:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's good:** <!-- one line -->\n\n### good-2\n\n**Company:**\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\nSubreddit:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's good:**\n\n<!-- Repeat good-3 \u2026 good-10 (target 10, spread across >3 companies). -->\n\n---\n\n## BAD examples\n\n### bad-1\n\n**Company:**\n**Company context:**\n\n```text\nCompany Name:\nCompany Description:\nIndustry / Sector:\nEmployee Count Range:\nCompany Categories:\nKnowledge Base Signals:\nSubreddit:\n```\n\n**Post:**\n\n```text\n\n```\n\n**Why it's bad:** <!-- one line: reads like an ad / corporate tone / breaks subreddit norms / ungrounded claim -->\n\n<!-- Repeat bad-2 \u2026 bad-5 (target 3\u20135). -->\n",
+} as const;
+
+export type ReferencePlatform = keyof typeof REFERENCE_POSTS;
